requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
pypdf2>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
groq>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.21.0
pandas>=2.0.0 